86cym3xfh: Rename module to property-listing from draft-listing
