+++ ui-modules/property-listing/components/ImageCarousel.tsx
